refactor: chatting full empty 병합
